pyproject.toml
[tool.poetry]
authors = ["Andrei Aksionau"]
description = "U-Net implementation with PyTorch"
name = "unet"
packages = [
  {include = "src"},
]
version = "1.0.0"

[tool.poetry.dependencies]
Pillow = "~9.4.0"
albumentations = "~1.3.0"
loguru = "~0.6.0"
numpy = "~1.24.1"
python = "~3.10.8"
torch = [
  {version = "~1.12.0+cpu", platform = "darwin"},
  {version = "~1.12.0", platform = "linux"},
]
torchvision = "~0.12.0"
tqdm = "~4.64.0"

[tool.poetry.group.dev.dependencies]
black = {version = "*", extras = ["jupyter"]}
flake8 = "*"
flake8-absolute-import = "*"
flake8-annotations = "*"
flake8-bugbear = "*"
flake8-commas = "*"
flake8-comprehensions = "*"
flake8-docstrings = "*"
flake8-eradicate = "*"
flake8-expression-complexity = "*"
flake8-isort = "*"
flake8-print = "*"
flake8-pytest-style = "*"

[tool.isort]
profile = "black"

[tool.black]
line_length = 120

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "pip>=21.0"]
